+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     - drop Handlebars dependency
 - Themes are now entrypoint-based [#829](https://github.com/opendatateam/udata/pull/829).
   There is also a new [theming documention](https://udata.readthedocs.io/en/stable/creating-theme/).
+- Images are now optimized and you can force rerendering using the `udata images render` command.
 
 ## 1.0.5 (2017-03-27)