+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   [#745](https://github.com/opendatateam/udata/issues/745)
 - Prevent a redirect on discussion fetch
   [#795](https://github.com/opendatateam/udata/issues/795)
+- API exposes both original and biggest thumbnail for organization logo, reuse image and user avatar
+  [#824](https://github.com/opendatateam/udata/issues/824)
 
 ## 1.0.5 (2017-03-27)